@@ -1,5 +1,5 @@
 /**
- * Copyright 2019, SumUp Ltd.
+ * Copyright 2025, SumUp Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,8 +13,25 @@
  * limitations under the License.
  */
 
-module.exports = require('./dist/eslint')({
-  rules: {
-    'no-process-exit': 'off',
+import type { Linter } from 'eslint';
+import compat from 'eslint-plugin-compat';
+import globals from 'globals';
+
+import { files } from './files.js';
+
+export const browser = {
+  name: 'foundry/browser',
+  files: [...files.javascript, ...files.typescript],
+  languageOptions: {
+    globals: globals.browser,
   },
-});
+  settings: {
+    lintAllEsApis: true,
+    // This API produces a false positive
+    polyfills: ['document.body'],
+  },
+  plugins: { compat },
+  rules: {
+    'compat/compat': 'error',
+  },
+} satisfies Linter.Config;
