@@ -19,13 +19,16 @@ import type { File } from '../../types/shared.js';
 
 export const files = (): File[] => [
   {
-    name: '.stylelintrc.js',
-    content: `
-    module.exports = require('@sumup-oss/foundry/stylelint')()`,
+    name: 'stylelint.config.mjs',
+    content: dedent`
+      import { defineConfig } from '@sumup-oss/foundry/stylelint';
+
+      export default defineConfig();
+    `,
   },
   {
     name: '.stylelintignore',
-    content: `${dedent`
+    content: dedent`
       node_modules/
       build/
       dist/
@@ -36,6 +39,6 @@ export const files = (): File[] => [
       coverage/
       __coverage__/
       __reports__/
-    `}\n`,
+    `,
   },
 ];
