@@ -15,16 +15,16 @@
 
 import { describe, expect, it, type Mock, vi } from 'vitest';
 
-import { getOptions as getOptionsMock } from '../../lib/options.js';
+import { getPlugins as getPluginsMock } from '../../lib/options.js';
 import { Plugin } from '../../types/shared.js';
 
-import { createConfig, customizeConfig } from './config.js';
+import { customizeConfig, defineConfig } from './config.js';
 
 vi.mock('../../lib/options', () => ({
-  getOptions: vi.fn(() => ({})),
+  getPlugins: vi.fn(() => []),
 }));
 
-const getOptions = getOptionsMock as Mock;
+const getPlugins = getPluginsMock as Mock;
 
 describe('stylelint', () => {
   describe('customizeConfig', () => {
@@ -104,9 +104,10 @@ describe('stylelint', () => {
       const overrides = {
         extends: ['stylelint-config-styled-components'],
       };
-      const actual = createConfig(overrides);
+      const actual = defineConfig(overrides);
       expect(actual).toEqual(
         expect.objectContaining({
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
           extends: expect.arrayContaining([
             'stylelint-config-standard',
             'stylelint-config-styled-components',
@@ -118,14 +119,14 @@ describe('stylelint', () => {
 
   describe('with options', () => {
     it("should return a config for { plugins: ['Circuit UI'] }", () => {
-      getOptions.mockReturnValue({ plugins: [Plugin.CIRCUIT_UI] });
-      const actual = createConfig();
+      getPlugins.mockReturnValue([Plugin.CircuitUI]);
+      const actual = defineConfig();
       expect(actual).toMatchSnapshot();
     });
 
     it("should return a config for { plugins: ['Circuit UI (OSS scope)'] }", () => {
-      getOptions.mockReturnValue({ plugins: [Plugin.CIRCUIT_UI_OSS] });
-      const actual = createConfig();
+      getPlugins.mockReturnValue([Plugin.CircuitUIOSS]);
+      const actual = defineConfig();
       expect(actual).toMatchSnapshot();
     });
   });
